requirements.txt
<<<<<<< HEAD
ultralytics>=8.0.0
opencv-python
numpy<2
pandas
torch==2.1.0
torchvision==0.16.0
matplotlib
flask
flask-cors
face_recognition
dlib
fastapi
uvicorn
python-multipart
pillow
openpyxl
psycopg2-binary
sqlalchemy
supabase==2.0.2
python-dotenv==1.0.0
=======
Flask==2.3.3
>>>>>>> 00201ebc57f662460ac3b5db0eea06552f371f65
